requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.7.4
filelock==3.18.0
Flask==3.1.3
flatbuffers==25.9.23
fonttools==4.60.2
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.16.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
InstructorEmbedding==1.0.1
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.7
langchain==0.2.3
langchain-community==0.2.3
langchain-core==0.2.43
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langcodes==3.5.0
langsmith==0.1.147
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.2
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.44.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
openai==1.93.0
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.17.0
orjson==3.10.18
packaging==23.2
pandas==2.3.0
parso==0.8.5
pillow==10.4.0
plotly==6.3.1
polars==1.36.1
polars-runtime-32==1.36.1
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.2.2
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
streamlit==1.35.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.29.2
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.4.37
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
waitress==3.0.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.14
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
yt-dlp==2025.10.14
zipp==3.23.0
zstandard==0.23.0
